+++ bittytax/conv/parsers/tradesatoshi.py
@@ -1,13 +1,16 @@
 # -*- coding: utf-8 -*-
 # (c) Nano Nano Ltd 2019
 
-from decimal import Decimal
+from decimal import Decimal, ROUND_DOWN
+
 
 from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
 from ..exceptions import UnexpectedTypeError
 
 WALLET = "TradeSatoshi"
+
+PRECISION = Decimal('0.00000000')
 
 def parse_tradesatoshi_deposits2(data_row, _parser, _filename):
     in_row = data_row.in_row
@@ -52,26 +55,30 @@ def parse_tradesatoshi_withdrawals(data_row, _parser, _filename):
                                              fee_asset=in_row[2],
                                              wallet=WALLET)
 
-def parse_tradesatoshi_trades(data_row, _parser, _filename):
+def parse_tradesatoshi_trades(data_row, parser, _filename):
     in_row = data_row.in_row
     data_row.timestamp = DataParser.parse_timestamp(in_row[6])
 
     if in_row[2] == "Buy":
+        sell_quantity = Decimal(in_row[3]) * Decimal(in_row[4])
+
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
                                                  data_row.timestamp,
                                                  buy_quantity=in_row[3],
                                                  buy_asset=in_row[1].split('/')[0],
-                                                 sell_quantity=Decimal(in_row[3]) * \
-                                                               Decimal(in_row[4]),
+                                                 sell_quantity=sell_quantity. \
+                                                         quantize(PRECISION, rounding=ROUND_DOWN),
                                                  sell_asset=in_row[1].split('/')[1],
                                                  fee_quantity=in_row[5],
                                                  fee_asset=in_row[1].split('/')[1],
                                                  wallet=WALLET)
     elif in_row[2] == "Sell":
+        buy_quantity = Decimal(in_row[3]) * Decimal(in_row[4])
+
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
                                                  data_row.timestamp,
-                                                 buy_quantity=Decimal(in_row[3]) * \
-                                                              Decimal(in_row[4]),
+                                                 buy_quantity=buy_quantity. \
+                                                         quantize(PRECISION, rounding=ROUND_DOWN),
                                                  buy_asset=in_row[1].split('/')[1],
                                                  sell_quantity=in_row[3],
                                                  sell_asset=in_row[1].split('/')[0],