+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Cryptopia parser: calculations rounded to 8 decimal places.
 - Tqdm workaround (https://github.com/tqdm/tqdm/issues/777).
 - Ledger Live parser: unrecognised operation type 'IN'.
+- TradeSatoshi parser: calculations rounded to 8 decimal places.
 ### Added
 - Conversion tool: added parser for CoinTracking.info accounting data.
 - Conversion tool: added parser for Gravity (Bitstocks) exchange.